widget.base: Add template support.
This allows some widgets to have declarative templates as first level
elements. It can be used for lazy loading or conditional loading.

This commit also improve get_children_by_id, but does not fix all the
corner case. As proposed by psychon, this will require them to be
moved to the hierarchy.